Split built-in agent catalog
