+++ apps/desktop/src/main/capability-tool-discovery.ts
@@ -9,7 +9,7 @@ import { StdioClientTransport } from '@modelcontextprotocol/sdk/client/stdio'
 import { StreamableHTTPClientTransport } from '@modelcontextprotocol/sdk/client/streamableHttp'
 import { getEffectiveSettings } from './settings.ts'
 import { getCustomAgentCatalog } from './custom-agents.ts'
-import { listBuiltInAgentDetails } from './agent-config.ts'
+import { listBuiltInAgentDetails } from './built-in-agent-details.ts'
 import { humanizeToolId, isVisibleRuntimeToolId, runtimeToolId } from './runtime-tools.ts'
 import { validateCustomMcpStdioCommand } from './mcp-stdio-policy.ts'
 import { listCustomMcps } from './native-customizations.ts'
