+++ tests/agent-config.test.ts
@@ -3,7 +3,8 @@ import assert from 'node:assert/strict'
 import { mkdtempSync, mkdirSync, rmSync, writeFileSync } from 'fs'
 import { join } from 'path'
 import { clearConfigCaches } from '../apps/desktop/src/main/config-loader.ts'
-import { buildOpenCoworkAgentConfig, listBuiltInAgentDetails } from '../apps/desktop/src/main/agent-config.ts'
+import { buildOpenCoworkAgentConfig } from '../apps/desktop/src/main/agent-config.ts'
+import { listBuiltInAgentDetails } from '../apps/desktop/src/main/built-in-agent-details.ts'
 
 function testTempDir(prefix: string) {
   const parent = join(process.cwd(), '.open-cowork-test')
